hr details CTA done hr details Delete done Deal list API data binding Deal Details HTMl integration with React.

+++ src/modules/hiring request/screens/hrdetail/hrdetails.js
@@ -29,6 +29,7 @@ import {
 import { MasterDAO } from 'core/master/masterDAO';
 import { UserSessionManagementController } from 'modules/user/services/user_session_services';
 import { hrUtils } from 'modules/hiring request/hrUtils';
+import { _isNull } from 'shared/utils/basic_utils';
 
 // import MatchmakingModal from 'modules/hiring request/components/matchmaking/matchmaking';
 
@@ -46,9 +47,6 @@ const ActivityFeed = React.lazy(() =>
 	import('modules/hiring request/components/activityFeed/activityFeed'),
 );
 
-const MatchmakingModal = React.lazy(() =>
-	import('modules/hiring request/components/matchmaking/matchmaking'),
-);
 const HRDetailScreen = () => {
 	const [deleteModal, setDeleteModal] = useState(false);
 	const [isLoading, setLoading] = useState(false);
@@ -56,7 +54,7 @@ const HRDetailScreen = () => {
 	const navigate = useNavigate();
 	const switchLocation = useLocation();
 	const [deleteReason, setDeleteReason] = useState([]);
-	const [activeTab, setActiveTab] = useState('');
+
 	const [adHOC, setAdHOC] = useState([
 		{
 			label: 'Pass to Pool',
@@ -76,10 +74,8 @@ const HRDetailScreen = () => {
 		register,
 		handleSubmit,
 		setValue,
-		control,
-		setError,
-		getValues,
 		watch,
+		setError,
 		formState: { errors },
 	} = useForm();
 
@@ -100,31 +96,42 @@ const HRDetailScreen = () => {
 		},
 		[navigate],
 	);
-	console.log(apiData, '--apiData--');
+
+	const clientOnLossSubmitHandler = useCallback(
+		async (d) => {
+			if (_isNull(watch('hrDeleteLossReason'))) {
+				setError('hrDeleteLossReason', 'Please select loss reason.');
+			} else if (_isNull(watch('hrDeleteLossRemark'))) {
+				setError('hrDeleteLossRemark', 'Please enter loss remark');
+			} else {
+				let deleteObj = {
+					id: urlSplitter?.split('HR')[0],
+					deleteType: HRDeleteType.LOSS,
+					reasonId: watch('hrDeleteLossReason').id,
+					otherReason: _isNull(watch('hrLossDeleteOtherReason'))
+						? ''
+						: watch('hrLossDeleteOtherReason'),
+					reason: watch('hrDeleteLossReason').value,
+					remark: watch('hrDeleteLossRemark'),
+					onBoardId: 0,
+				};
+
+				let deletedResponse = await hiringRequestDAO.deleteHRDAO(deleteObj);
+				if (
+					deletedResponse &&
+					deletedResponse.statusCode === HTTPStatusCode.OK
+				) {
+					navigate(UTSRoutes.ALLHIRINGREQUESTROUTE);
+				}
+			}
+		},
+		[navigate, setError, urlSplitter, watch],
+	);
 	const clientOnHoldSubmitHandler = useCallback(
 		async (d) => {
 			let deleteObj = {
 				id: urlSplitter?.split('HR')[0],
 				deleteType: HRDeleteType.ON_HOLD,
-				reasonId: d.hrDeleteReason.id,
-				otherReason: d.hrDeleteOtherReason,
-				reason: d.hrDeleteReason.value,
-				remark: d.hrDeleteRemark,
-				onBoardId: 0,
-			};
-
-			let deletedResponse = await hiringRequestDAO.deleteHRDAO(deleteObj);
-			if (deletedResponse && deletedResponse.statusCode === HTTPStatusCode.OK) {
-				navigate(UTSRoutes.ALLHIRINGREQUESTROUTE);
-			}
-		},
-		[navigate, urlSplitter],
-	);
-	const clientLossSubmitHandler = useCallback(
-		async (d) => {
-			let deleteObj = {
-				id: urlSplitter?.split('HR')[0],
-				deleteType: HRDeleteType.LOSS,
 				reasonId: d.hrDeleteReason.id,
 				otherReason: d.hrDeleteOtherReason,
 				reason: d.hrDeleteReason.value,
@@ -251,7 +258,11 @@ const HRDetailScreen = () => {
 							<Skeleton active />
 						) : (
 							<Suspense>
-								<TalentProfileCard talentDetail={apiData?.HRTalentDetails} />
+								<TalentProfileCard
+									HRStatusCode={apiData?.HRStatusCode}
+									talentDetail={apiData?.HRTalentDetails}
+									miscData={miscData}
+								/>
 							</Suspense>
 						)}
 					</div>
@@ -270,171 +281,187 @@ const HRDetailScreen = () => {
 						</Suspense>
 					)}
 				</div>
-
-				{/** ------------------ HR Delete Modal ---------------------- */}
-				<Modal
-					transitionName=""
-					centered
-					open={deleteModal}
-					width={'864px'}
-					footer={null}
-					onCancel={() => setDeleteModal(false)}>
-					<div className={HRDetailStyle.modalBody}>
-						<div>
-							<h1>Delete Type</h1>
-							<p>{updatedSplitter}</p>
-						</div>
-
-						<div className={HRDetailStyle.tabContainer}>
-							<Tabs
-								defaultActiveKey="1"
-								animated={true}
-								centered
-								tabBarStyle={{
-									borderBottom: `1px solid var(--uplers-border-color)`,
-								}}
-								items={[
-									{
-										label: 'On Hold',
-										key: 'On Hold',
-										children: (
-											<div style={{ marginTop: '35px' }}>
-												<div className={HRDetailStyle.row}>
-													<div className={HRDetailStyle.colMd12}>
-														<HRSelectField
-															mode={'id/value'}
-															searchable={false}
-															setValue={setValue}
-															register={register}
-															name="hrDeleteReason"
-															label="Delete Reason"
-															defaultValue="Please select reason"
-															options={deleteReason && deleteReason}
-															required
-															isError={
-																errors['hrDeleteReason'] &&
-																errors['hrDeleteReason']
-															}
-															errorMsg="Please select a delete reason."
-														/>
-													</div>
-												</div>
-												{watch('hrDeleteReason')?.value === 'Other' && (
-													<div className={HRDetailStyle.row}>
-														<div className={HRDetailStyle.colMd12}>
-															<HRInputField
-																register={register}
-																errors={errors}
-																validationSchema={{
-																	required: 'Please enter the other reason.',
-																}}
-																label="On Hold Other Reason"
-																name="hrDeleteOtherReason"
-																type={InputType.TEXT}
-																placeholder="Enter Other Reason"
-																required
-															/>
-														</div>
-													</div>
-												)}
-												<div className={HRDetailStyle.row}>
-													<div className={HRDetailStyle.colMd12}>
-														<HRInputField
-															isTextArea={true}
-															register={register}
-															errors={errors}
-															validationSchema={{
-																required: 'Please enter the HR Remark.',
-															}}
-															label="On Hold Remark"
-															name="hrDeleteRemark"
-															type={InputType.TEXT}
-															placeholder="Enter Remark"
-															required
-														/>
-													</div>
-												</div>
-												<div className={HRDetailStyle.formPanelAction}>
-													<button
-														onClick={() => setDeleteModal(false)}
-														className={HRDetailStyle.btn}>
-														Cancel
-													</button>
-													<button
-														id={HRDeleteType.ON_HOLD}
-														type="submit"
-														onClick={handleSubmit(clientOnHoldSubmitHandler)}
-														className={HRDetailStyle.btnPrimary}>
-														Delete
-													</button>
-												</div>
-											</div>
-										),
-									},
-									{
-										label: 'Loss',
-										key: 'Loss',
-										children: (
-											<div style={{ marginTop: '35px' }}>
-												<div className={HRDetailStyle.row}>
-													<div className={HRDetailStyle.colMd12}>
-														<HRSelectField
-															searchable={false}
-															setValue={setValue}
-															register={register}
-															name="hrDeleteLossReason"
-															label="Delete Reason"
-															defaultValue="Please select reason"
-															options={deleteReason && deleteReason}
-															required
-															isError={
-																errors['hrDeleteLossReason'] &&
-																errors['hrDeleteLossReason']
-															}
-															errorMsg="Please select a delete reason."
-														/>
-													</div>
-												</div>
-												<div className={HRDetailStyle.row}>
-													<div className={HRDetailStyle.colMd12}>
-														<HRInputField
-															isTextArea={true}
-															register={register}
-															errors={errors}
-															validationSchema={{
-																required: 'please enter the HR Remark.',
-															}}
-															label="Loss Remark"
-															name="hrDeleteLossRemark"
-															type={InputType.TEXT}
-															placeholder="Enter Remark"
-															required
-														/>
-													</div>
-												</div>
-												<div className={HRDetailStyle.formPanelAction}>
-													<button
-														onClick={() => setDeleteModal(false)}
-														className={HRDetailStyle.btn}>
-														Cancel
-													</button>
-													<button
-														id={HRDeleteType.LOSS}
-														type="submit"
-														// onClick={handleSubmit(clientLossSubmitHandler)}
-														className={HRDetailStyle.btnPrimary}>
-														Delete
-													</button>
-												</div>
-											</div>
-										),
-									},
-								]}
-							/>
-						</div>
-					</div>
-				</Modal>
 			</div>
+			{/** ------------------ HR Delete Modal ---------------------- */}
+			<Modal
+				transitionName=""
+				centered
+				open={deleteModal}
+				width={'864px'}
+				footer={null}
+				onCancel={() => setDeleteModal(false)}>
+				<div className={HRDetailStyle.modalBody}>
+					<div>
+						<h1>Delete Type</h1>
+						<p>{updatedSplitter}</p>
+					</div>
+
+					<div className={HRDetailStyle.tabContainer}>
+						<Tabs
+							defaultActiveKey="1"
+							animated={true}
+							centered
+							tabBarStyle={{
+								borderBottom: `1px solid var(--uplers-border-color)`,
+							}}
+							items={[
+								{
+									label: 'On Hold',
+									key: 'On Hold',
+									children: (
+										<div style={{ marginTop: '35px' }}>
+											<div className={HRDetailStyle.row}>
+												<div className={HRDetailStyle.colMd12}>
+													<HRSelectField
+														mode={'id/value'}
+														searchable={false}
+														setValue={setValue}
+														register={register}
+														name="hrDeleteReason"
+														label="Delete Reason"
+														defaultValue="Please select reason"
+														options={deleteReason && deleteReason}
+														required
+														isError={
+															errors['hrDeleteReason'] &&
+															errors['hrDeleteReason']
+														}
+														errorMsg="Please select a delete reason."
+													/>
+												</div>
+											</div>
+											{watch('hrDeleteReason')?.value === 'Other' && (
+												<div className={HRDetailStyle.row}>
+													<div className={HRDetailStyle.colMd12}>
+														<HRInputField
+															register={register}
+															errors={errors}
+															validationSchema={{
+																required: 'Please enter the other reason.',
+															}}
+															label="On Hold Other Reason"
+															name="hrDeleteOtherReason"
+															type={InputType.TEXT}
+															placeholder="Enter Other Reason"
+															required
+														/>
+													</div>
+												</div>
+											)}
+											<div className={HRDetailStyle.row}>
+												<div className={HRDetailStyle.colMd12}>
+													<HRInputField
+														isTextArea={true}
+														register={register}
+														errors={errors}
+														validationSchema={{
+															required: 'Please enter the HR Remark.',
+														}}
+														label="On Hold Remark"
+														name="hrDeleteRemark"
+														type={InputType.TEXT}
+														placeholder="Enter Remark"
+														required
+													/>
+												</div>
+											</div>
+											<div className={HRDetailStyle.formPanelAction}>
+												<button
+													onClick={() => setDeleteModal(false)}
+													className={HRDetailStyle.btn}>
+													Cancel
+												</button>
+												<button
+													id={HRDeleteType.ON_HOLD}
+													type="submit"
+													onClick={handleSubmit(clientOnHoldSubmitHandler)}
+													className={HRDetailStyle.btnPrimary}>
+													Delete
+												</button>
+											</div>
+										</div>
+									),
+								},
+								{
+									label: 'Loss',
+									key: 'Loss',
+									children: (
+										<div style={{ marginTop: '35px' }}>
+											<div className={HRDetailStyle.row}>
+												<div className={HRDetailStyle.colMd12}>
+													<HRSelectField
+														mode={'id/value'}
+														searchable={false}
+														setValue={setValue}
+														register={register}
+														name="hrDeleteLossReason"
+														label="Delete Reason"
+														defaultValue="Please select reason"
+														options={deleteReason && deleteReason}
+														required
+														isError={
+															errors['hrDeleteLossReason'] &&
+															errors['hrDeleteLossReason']
+														}
+														errorMsg="Please select a delete reason."
+													/>
+												</div>
+											</div>
+											{watch('hrDeleteLossReason')?.value === 'Other' && (
+												<div className={HRDetailStyle.row}>
+													<div className={HRDetailStyle.colMd12}>
+														<HRInputField
+															register={register}
+															errors={errors}
+															validationSchema={{
+																required: 'Please enter the other reason.',
+															}}
+															label="On Loss Other Reason"
+															name="hrLossDeleteOtherReason"
+															type={InputType.TEXT}
+															placeholder="Enter Other Reason"
+															required
+														/>
+													</div>
+												</div>
+											)}
+											<div className={HRDetailStyle.row}>
+												<div className={HRDetailStyle.colMd12}>
+													<HRInputField
+														isTextArea={true}
+														register={register}
+														errors={errors}
+														validationSchema={{
+															required: 'please enter the HR Remark.',
+														}}
+														label="Loss Remark"
+														name="hrDeleteLossRemark"
+														type={InputType.TEXT}
+														placeholder="Enter Remark"
+														required
+													/>
+												</div>
+											</div>
+											<div className={HRDetailStyle.formPanelAction}>
+												<button
+													onClick={() => setDeleteModal(false)}
+													className={HRDetailStyle.btn}>
+													Cancel
+												</button>
+												<button
+													onClick={clientOnLossSubmitHandler}
+													className={HRDetailStyle.btnPrimary}>
+													Delete
+												</button>
+											</div>
+										</div>
+									),
+								},
+							]}
+						/>
+					</div>
+				</div>
+			</Modal>
 		</WithLoader>
 	);
 };
